backend/python-services/speech_to_text/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
openai-whisper==20231117
numpy>=1.24.0
torch>=2.1.0
pydantic==2.6.0
